Use `yarn add` with a path instead of `yarn link`
Both `yarn link` and `npm link` do not install the dependencies
of the linked package. This means that things like leaflet are
not actually available in a dev/test context when building the
frontend, even though they would be in a production context
where the frontend package is installed normally.

This PR fixes the situation by borrowing the same approach as
blacklight, which instead adds the package with `yarn add` pointing
to the parent directory. This way, its dependencies will be installed
at the correct level and available for building with Vite.